refactor: migrate init element tree v1 to v2 continued
