--- libs/frontend/modules/builder/src/graphql/index.ts
@@ -1,1 +1,0 @@
-export * from './page-builder.endpoints.graphql.gen'
