+++ libs/frontend/modules/builder/src/index.ts
@@ -1,5 +1,4 @@
 export * from './dnd'
-export * from './graphql'
 export * from './hooks'
 export * from './sections'
 export * from './sections/content/Builder'
