--- libs/frontend/modules/builder/src/store/builder.api.ts
@@ -1,4 +1,0 @@
-import { client } from '@codelab/frontend/model/infra/graphql'
-import { getSdk } from '../graphql'
-
-export const pageBuilderApi = getSdk(client)
