+++ libs/frontend/modules/builder/src/store/index.ts
@@ -1,2 +1,1 @@
-export * from './builder.api'
 export * from './builder.service'
