+++ libs/shared/abstract/core/src/domain/component/component.model.interface.ts
@@ -4,6 +4,7 @@ import { ICacheService } from '../../service'
 import { IElementTreeService } from '../element'
 import { IAnyType } from '../type'
 import { IComponentDTO } from './component.dto.interface'
+import { RenderedComponentFragment } from './component-render.fragment.graphql.gen'
 
 export interface IComponent
   extends INodeType<'Component'>,
@@ -14,6 +15,7 @@ export interface IComponent
   rootElementId: string
   ownerId: string
   api: Ref<IAnyType>
+  loadComponentTree(renderedComponentFragment: RenderedComponentFragment): void
 }
 
 export type IComponentRef = string
